requirements.txt
stealth_requests>=1.0.0
discord-webhook>=1.0.0
beautifulsoup4>=4.9.3
aiohttp>=3.8.1
python-dotenv>=0.19.0